+++ src/translators/openai.response.message.translators.ts
@@ -1,8 +1,8 @@
 import 'reflect-metadata';
 import {injectable} from 'tsyringe';
-import {createStableId, pickDefined, safeParse} from "@holokai/sdk";
-import {HoloContent, HoloMessage} from "@holokai/types/holo";
-import {BaseTranslator} from "@holokai/sdk/provider";
+import {createStableId, pickDefined, safeParse} from "@holokai/holo-sdk";
+import {HoloContent, HoloMessage} from "@holokai/holo-types/holo";
+import {BaseTranslator} from "@holokai/holo-sdk/provider";
 import {ChatCompletionMessage} from "openai/resources/chat/completions/completions";
 
 @injectable()