+++ src/translators/streaming/openai.stream.translator.ts
@@ -4,8 +4,8 @@ import {OpenAIMessageStartTranslator} from './openai.message.start.translator';
 import {OpenAIContentDeltaTranslator} from './openai.content.delta.translator';
 import {OpenAIMessageDeltaTranslator} from './openai.message.delta.translator';
 import {OpenAIMessageStopTranslator} from './openai.message.stop.translator';
-import {HoloStreamChunk} from "@holokai/types/holo";
-import {StreamTranslator} from "@holokai/sdk/provider";
+import {HoloStreamChunk} from "@holokai/holo-types/holo";
+import {StreamTranslator} from "@holokai/holo-sdk/provider";
 import {ChatCompletionChunk} from "openai/resources/chat/completions/completions";
 
 @injectable()