+++ CHANGELOG.md
@@ -70,17 +70,17 @@
 
 - fd64cb1: Moved app to separate workspace so that we can track with changeset.
 - Updated dependencies
-  - @holokai/sdk@1.1.0
+  - @holokai/holo-sdk@1.1.0
 
 ## 1.0.0
 
 ### Minor Changes
 
-- Broke up SDK to have dependency-less @holokai/types. Reorganized types and interfaces. Fixed up some types vs
+- Broke up SDK to have dependency-less @holokai/holo-types. Reorganized types and interfaces. Fixed up some types vs
   interfaces. Prefixed true interfaces that are used to describe classes vs TypeScript interfaces used to just describe
   an object.
 
 ### Patch Changes
 
 - Updated dependencies
-  - @holokai/sdk@1.0.0
+  - @holokai/holo-sdk@1.0.0